Properly use VueX store
