+++ app/client/src/vue/components/message/SingleMessage.vue
@@ -41,6 +41,8 @@
 
 <script>
     import Vue from "vue";
+    import { mapState } from 'vuex';
+    
     export default Vue.extend({
         props: {
             message: {
