+++ app/client/src/vue/components/overlay/CursorOverlay.vue
@@ -1,7 +1,7 @@
 <template>
-    <div class="cursor-container" v-show="showCursors">
+    <div class="cursor-container" v-show="cursor">
         <div class="cursor"
-             v-for="entry in cursors"
+             v-for="entry in cursorList"
              :key="entry.cursor.user.id"
              :style="{left: (entry.cursor.x * window.innerWidth) + 'px', top: (entry.cursor.y * window.innerHeight) + 'px'}">
             <img src="/assets/images/cursors/default.png">
@@ -12,6 +12,7 @@
 
 <script>
     import Vue from "vue";
+    import { mapState } from 'vuex';
 
     const CURSOR_POSITION_DELAY_MS = 150;
 
@@ -36,8 +37,8 @@
             document.addEventListener('mousemove', event => {
                 const x = event.clientX / window.innerWidth;
                 const y = event.clientY / window.innerHeight;
-                if (this.$store.state.currentRoom !== null
-                    && this.$store.state && this.$store.state.user.data.plugins.cursor) {
+                if (this.$store.state.Main.currentRoom !== null
+                    && this.$store.state && this.$store.state.Main.user.data.plugins.cursor) {
                     this.sendCursorPosition(x, y);
                 }
             });
@@ -56,12 +57,10 @@
             }
         },
         computed: {
-            cursors: function() {
-                return Object.values(this.$store.state.cursors);
-            },
-            showCursors: function() {
-                return this.$store.state.user.data.plugins.cursor;
-            }
+            ...mapState('Main', {
+                'cursorList': state => Object.values(state.cursors),
+                'user': state => state.user.data.plugins.cursor,
+            }),
         }
     });
 </script>
