+++ app/client/src/vue/components/util/HoverCard.vue
@@ -15,9 +15,7 @@
     import Vue from "vue";
 
     export default Vue.extend({
-        
         props: ['selected', 'highlighted', 'borderColor'],
-
     });
 </script>
