+++ app/client/src/vue/components/video-player/implementations/YoutubePlayer.vue
@@ -11,6 +11,7 @@
 
 <script>
     import Vue from "vue";
+    import { mapState } from "vuex";
 
     export default Vue.extend({
         data: function() { return { src: '', previousVideoHash: null, } },
@@ -38,12 +39,10 @@
             }
         },
         computed: {
-            playerState: function() {
-                return this.$store.state.playerState;
-            },
-            playerStateLastUpdate: function() {
-                return this.$store.state.playerStateLastUpdate;
-            },
+            ...mapState('Main', [
+                'playerState',
+                'playerStateLastUpdate',
+            ]),
         }
     });
 </script>
