+++ app/client/src/vue/store/index.js
@@ -1,0 +1,3 @@
+import store from './store';
+
+export default store;
