+++ webpack.config.js
@@ -5,7 +5,6 @@ module.exports = {
     output: {
         filename: 'bundle.js'
     },
-    watch: true,
     module: {
         rules: [
             {
